projects/python/9.fastapi-testing/requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
alembic
pytest
httpx